refactor: move pagination to core
