+++ packages/core/src/index.ts
@@ -10,5 +10,8 @@ export { Silence } from '@stops/Silence';
 
 export { CustomId } from '@customId/CustomId';
 
+export { paginate } from '@pagination/paginate';
+export { type PageView } from '@pagination/PageView';
+
 /** Package version */
 export const version = process.env.PACKAGE_VERSION ?? '0.0.0';
