+++ packages/core/src/internal.index.ts
@@ -4,4 +4,6 @@ export { NoticeCard } from '@stops/NoticeCard';
 
 export { prefixOf, decodeFor, type AnyCustomId } from '@customId/CustomId';
 export { ComponentDefsKey, type HasComponentDefs } from '@customId/routing';
-export type { CustomIdField, DecodedParams } from '@customId/Field';
+export type { DecodedParams } from '@customId/Field';
+
+export { PAGE_MAX, pageCursor, type PageCursor } from '@pagination/cursor';
