+++ packages/core/tsconfig.json
@@ -8,6 +8,7 @@
             "@src/*": ["./src/*"],
             "@components/*": ["./src/components/*"],
             "@customId/*": ["./src/customId/*"],
+            "@pagination/*": ["./src/pagination/*"],
             "@registries/*": ["./src/registries/*"],
             "@stops/*": ["./src/stops/*"]
         }
