+++ packages/kit/src/index.ts
@@ -1,5 +1,2 @@
-export { paginate } from '@pagination/paginate';
-export { type PageView } from '@pagination/PageView';
-
 /** Package version */
 export const version = process.env.PACKAGE_VERSION ?? '0.0.0';
