+++ packages/kit/src/internal.index.ts
@@ -1,1 +1,1 @@
-export { PAGE_MAX, pageCursor, type PageCursor } from '@pagination/cursor';
+export {};
